requirements.txt
requests
tqdm==4.64.0
pandas==1.4.3
biopython==1.79
streamlit==1.12.0
numpy==1.23.2
textblob
aiohttp
